feat: renamed files, new TableOfContentsAnchor component
