+++ packages/paste-website/src/components/shortcodes/table-of-contents/TableOfContentsAnchor.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import {Box} from '@twilio-paste/box';
+import type {TableOfContentsAnchorProps} from './types';
+
+export const TableOfContentsAnchor: React.FC<TableOfContentsAnchorProps> = ({href, children, ...props}) => {
+  return (
+    <Box
+      {...props}
+      as="a"
+      href={href}
+      data-cy="page-aside-anchor"
+      color="inherit"
+      textDecoration="none"
+      display="inline-block"
+    >
+      {children}
+    </Box>
+  );
+};
