+++ packages/paste-website/src/components/shortcodes/table-of-contents/types.ts
@@ -1,3 +1,5 @@
+import type {BoxStyleProps} from '@twilio-paste/box';
+
 export interface TableOfContentsProps {
   headings?: [
     {
@@ -11,4 +13,9 @@ export interface TableOfContentsListItemProps {
   active?: boolean;
   depth?: string;
   key?: string;
+}
+
+export interface TableOfContentsAnchorProps extends BoxStyleProps {
+  href: string;
+  children: string;
 }
